Add more documentation and refactor EclipseMarkupModelListener
