+++ core/src/nl/tudelft/watchdog/core/logic/event/eventtypes/staticanalysis/FileWarningSnapshotEvent.java
@@ -11,7 +11,8 @@ import java.util.List;
 
 /**
  * Analog to a {@link StaticAnalysisWarningEvent}, but this time it stores a list of {@link Warning},
- * as a snapshot of the current state of the {@link #document}.
+ * as a snapshot of the current state of the {@link #document}. If you need to save multiple events,
+ * use this event rather than {@link StaticAnalysisWarningEvent}.
  */
 public class FileWarningSnapshotEvent extends EventBase {
