+++ eclipse/plugin/src/nl/tudelft/watchdog/eclipse/ui/handlers/StartupUIThread.java
@@ -41,7 +41,7 @@ public class StartupUIThread implements Runnable {
 
 	@Override
 	public void run() {
-		// Do not open the window during tests. This would block the Thread
+		// Do not open the registration dialog during tests. This would block the Thread
 		// and not terminate on Travis
 		if (System.getProperty("suppressWatchdogRegistration") != null) {
 			return;
